+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/PerformanceCachingGoogleCloudStorage.java
@@ -13,16 +13,13 @@
  */
 package com.google.cloud.hadoop.gcsio;
 
-import static com.google.api.client.util.Strings.isNullOrEmpty;
-import static com.google.common.base.Strings.nullToEmpty;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
+import java.time.Duration;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
-import javax.annotation.Nullable;
 
 /**
  * This class adds a caching layer around a GoogleCloudStorage instance, caching calls that create,
@@ -37,12 +34,6 @@ public class PerformanceCachingGoogleCloudStorage extends ForwardingGoogleCloudS
   /** Cache to hold item info and manage invalidation. */
   private final PrefixMappedItemCache cache;
 
-  /** Options that configure delegate storage. */
-  private final GoogleCloudStorageOptions delegateOptions;
-
-  /** Options that configure this cache. */
-  private final PerformanceCachingGoogleCloudStorageOptions options;
-
   /**
    * Creates a wrapper around a GoogleCloudStorage instance, caching calls that create, update,
    * remove, and query for GoogleCloudStorageItemInfo. Those cached copies are returned when
@@ -56,25 +47,20 @@ public class PerformanceCachingGoogleCloudStorage extends ForwardingGoogleCloudS
    */
   public PerformanceCachingGoogleCloudStorage(
       GoogleCloudStorage delegate, PerformanceCachingGoogleCloudStorageOptions options) {
-    this(delegate, options, createCache(options));
+    this(delegate, createCache(options));
   }
 
   @VisibleForTesting
   PerformanceCachingGoogleCloudStorage(
       GoogleCloudStorage delegate,
-      PerformanceCachingGoogleCloudStorageOptions options,
       PrefixMappedItemCache cache) {
     super(delegate);
-    this.delegateOptions = delegate.getOptions();
-    this.options = options;
     this.cache = cache;
   }
 
   private static PrefixMappedItemCache createCache(
       PerformanceCachingGoogleCloudStorageOptions options) {
-    PrefixMappedItemCache.Config config = new PrefixMappedItemCache.Config();
-    config.setMaxEntryAgeMillis(options.getMaxEntryAgeMillis());
-    return new PrefixMappedItemCache(config);
+    return new PrefixMappedItemCache(Duration.ofMillis(options.getMaxEntryAgeMillis()));
   }
 
   @Override
@@ -88,12 +74,12 @@ public class PerformanceCachingGoogleCloudStorage extends ForwardingGoogleCloudS
   }
 
   @Override
-  public void deleteObjects(List<StorageResourceId> fullObjectNames) throws IOException {
-    super.deleteObjects(fullObjectNames);
+  public void deleteObjects(List<StorageResourceId> resourceIds) throws IOException {
+    super.deleteObjects(resourceIds);
 
     // Remove the deleted objects from cache.
-    for (StorageResourceId id : fullObjectNames) {
-      cache.removeItem(id);
+    for (StorageResourceId resourceId : resourceIds) {
+      cache.removeItem(resourceId);
     }
   }
 
@@ -124,24 +110,9 @@ public class PerformanceCachingGoogleCloudStorage extends ForwardingGoogleCloudS
       throws IOException {
     List<GoogleCloudStorageItemInfo> result;
 
-    if (options.isListCachingEnabled()) {
-      result = cache.getList(bucketName, objectNamePrefix);
-
-      if (result == null) {
-        result = super.listObjectInfo(bucketName, objectNamePrefix, null);
-        cache.putList(bucketName, objectNamePrefix, result);
-      }
-
-      filter(result, bucketName, objectNamePrefix, delimiter);
-
-      if (maxResults > 0 && result.size() > maxResults) {
-        result = result.subList(0, (int) maxResults);
-      }
-    } else {
-      result = super.listObjectInfo(bucketName, objectNamePrefix, delimiter, maxResults);
-      for (GoogleCloudStorageItemInfo item : result) {
-        cache.putItem(item);
-      }
+    result = super.listObjectInfo(bucketName, objectNamePrefix, delimiter, maxResults);
+    for (GoogleCloudStorageItemInfo item : result) {
+      cache.putItem(item);
     }
 
     return result;
@@ -151,102 +122,12 @@ public class PerformanceCachingGoogleCloudStorage extends ForwardingGoogleCloudS
   public ListPage<GoogleCloudStorageItemInfo> listObjectInfoPage(
       String bucketName, String objectNamePrefix, String delimiter, String pageToken)
       throws IOException {
-    if (options.isListCachingEnabled()) {
-      return new ListPage<>(listObjectInfo(bucketName, objectNamePrefix, delimiter), null);
-    }
-
     ListPage<GoogleCloudStorageItemInfo> result =
         super.listObjectInfoPage(bucketName, objectNamePrefix, delimiter, pageToken);
     for (GoogleCloudStorageItemInfo item : result.getItems()) {
       cache.putItem(item);
     }
     return result;
-  }
-
-  /**
-   * Matches Google Cloud Storage's delimiter filtering.
-   *
-   * @param items the mutable list of items to filter. Items matching the filter conditions will be
-   *     removed from this list.
-   * @param bucketName the bucket name to filter for.
-   * @param prefix the object name prefix to filter for.
-   * @param delimiter the delimiter to filter on.
-   */
-  private void filter(
-      List<GoogleCloudStorageItemInfo> items,
-      String bucketName,
-      @Nullable String prefix,
-      @Nullable String delimiter)
-      throws IOException {
-    prefix = nullToEmpty(prefix);
-
-    // if delimiter is not specified we don't need to filter-out subdirectories
-    if (isNullOrEmpty(delimiter)) {
-      // if prefix is not specified it means that we are listing all objects in the bucket
-      // and we need to exclude bucket from the result
-      if (prefix.isEmpty()) {
-        Iterator<GoogleCloudStorageItemInfo> itr = items.iterator();
-        while (itr.hasNext()) {
-          GoogleCloudStorageItemInfo item = itr.next();
-          if (item.isBucket()) {
-            itr.remove();
-            break;
-          }
-        }
-      }
-      return;
-    }
-
-    HashSet<String> dirs = new HashSet<>();
-    Iterator<GoogleCloudStorageItemInfo> itr = items.iterator();
-    while (itr.hasNext()) {
-      GoogleCloudStorageItemInfo item = itr.next();
-      String objectName = item.getObjectName();
-
-      // 1. Remove if bucket (means that listing objects in bucket):
-      //    do not return bucket itself (prefix dir) to avoid infinite recursion
-      // 2. Remove if doesn't start with the prefix.
-      // 3. Remove prefix object if it ends with delimiter:
-      //    do not return prefix dir to avoid infinite recursion.
-      if (item.isBucket()
-          || !objectName.startsWith(prefix)
-          || (prefix.endsWith(delimiter) && objectName.equals(prefix))) {
-        itr.remove();
-      } else {
-        // Retain if missing the delimiter after the prefix.
-        int firstIndex = objectName.indexOf(delimiter, prefix.length());
-        if (firstIndex != -1) {
-          // Remove if the first occurrence of the delimiter after the prefix isn't the last.
-          // Remove if the last occurrence of the delimiter isn't the end of the string.
-          int lastIndex = objectName.lastIndexOf(delimiter);
-          if (firstIndex != lastIndex || lastIndex != objectName.length() - 1) {
-            itr.remove();
-            dirs.add(objectName.substring(0, firstIndex + 1));
-          }
-        }
-      }
-    }
-
-    // Remove non-implicit directories (i.e. have corresponding directory objects)
-    for (GoogleCloudStorageItemInfo item : items) {
-      dirs.remove(item.getObjectName());
-    }
-
-    if (dirs.isEmpty()) {
-      return;
-    }
-
-    List<StorageResourceId> dirIds = new ArrayList<>(dirs.size());
-    for (String dir : dirs) {
-      dirIds.add(new StorageResourceId(bucketName, dir));
-    }
-
-    boolean inferImplicitDirectories = delegateOptions.isInferImplicitDirectoriesEnabled();
-    if (inferImplicitDirectories) {
-      for (StorageResourceId dirId : dirIds) {
-        items.add(GoogleCloudStorageItemInfo.createInferredDirectory(dirId));
-      }
-    }
   }
 
   /** This function may return cached copies of GoogleCloudStorageItemInfo. */
@@ -263,16 +144,9 @@ public class PerformanceCachingGoogleCloudStorage extends ForwardingGoogleCloudS
       int lastSlashIndex = objectName.lastIndexOf(PATH_DELIMITER);
       String directoryName =
           lastSlashIndex >= 0 ? objectName.substring(0, lastSlashIndex + 1) : null;
-      List<GoogleCloudStorageItemInfo> cachedInDirectory =
-          cache.listItems(bucketName, directoryName);
-      filter(cachedInDirectory, bucketName, directoryName, PATH_DELIMITER);
-      // If there are items already cached in directory, do not prefetch with list requests,
-      // because metadata for this directory already could be prefetched
-      if (cachedInDirectory.isEmpty()) {
-        // make just 1 request to prefetch only 1 page of directory items
-        listObjectInfoPage(bucketName, directoryName, PATH_DELIMITER, /* pageToken= */ null);
-        item = cache.getItem(resourceId);
-      }
+      // make just 1 request to prefetch only 1 page of directory items
+      listObjectInfoPage(bucketName, directoryName, PATH_DELIMITER, /* pageToken= */ null);
+      item = cache.getItem(resourceId);
     }
 
     // If it wasn't in the cache and wasn't cached in directory list request