+++ gcs/CHANGES.md
@@ -32,6 +32,9 @@
 
 1.  Use Slf4j backend by default for Google Flogger.
 
+1.  Remove list requests caching in the `PerformanceCachingGoogleCloudStorage`
+    and `fs.gs.performance.cache.list.caching.enable` property.
+
 ### 2.0.1 - 2020-02-13
 
 1.  Cooperative Locking FSCK tool: fix recovery of operations that failed before